remove old TJLedSequence code
