+++ examples/sequence/sequence.ino
@@ -1,11 +1,11 @@
-// JLed multi LED demo. 'Play' multiple LEDs, one after another.
-// Copyright 2019 by Jan Delgado. All rights reserved.
+// JLed multi LED demo. 'Play' multiple LEDs, one after another using a group.
+// Copyright 2019-2026 by Jan Delgado. All rights reserved.
 // https://github.com/jandelgado/jled
 #include <jled.h>
 
 constexpr auto LED_PIN = 3;
 
-JLed leds[] = {
+JLedAny leds[] = {
     JLed(LED_PIN).Breathe(2000).Repeat(3),
     JLed(LED_PIN).Blink(750, 250).Repeat(3),
     JLed(LED_PIN).FadeOff(1000).Repeat(3),
@@ -14,7 +14,7 @@ JLed leds[] = {
     JLed(LED_PIN).Off()
 };
 
-JLedSequence sequence(JLedSequence::eMode::SEQUENCE, leds);
+auto sequence = JLedGroup::Sequential(leds);
 
 void setup() { }
