+++ platformio.ini
@@ -31,12 +31,12 @@ src_dir = examples/hello
 ;src_dir = examples/hires
 ;src_dir = examples/simple_on
 ;src_dir = examples/last_brightness
-;src_dir = examples/multiled
-;src_dir = examples/multiled_mbed
 ;src_dir = examples/user_func
 ;src_dir = examples/sequence
 ;src_dir = examples/group
+;src_dir = examples/nested_group
 ;src_dir = examples/group_ref
+;src_dir = examples/group_mbed
 ;src_dir = examples/custom_hal
 ;src_dir = examples/pulse
 ;src_dir = examples/fade_from_to
