+++ test/Makefile
@@ -20,9 +20,6 @@ TEST_ARDUINO_MOCK_OBJECTS=$(TEST_ARDUINO_MOCK_SRC:.cpp=.o)
 TEST_JLED_SRC=Arduino.cpp test_jled.cpp test_main.cpp ../src/jled_base.cpp ${CATCH}
 TEST_JLED_OBJECTS=$(TEST_JLED_SRC:.cpp=.o)
 
-TEST_JLED_SEQUENCE_SRC=Arduino.cpp test_jled_sequence.cpp test_main.cpp ../src/jled_base.cpp ${CATCH}
-TEST_JLED_SEQUENCE_OBJECTS=$(TEST_JLED_SEQUENCE_SRC:.cpp=.o)
-
 TEST_ESP32_SRC=esp-idf/esp_timer.cpp esp-idf/driver/ledc.cpp \
 			   test_esp32_hal.cpp test_main.cpp ${CATCH}
 TEST_ESP32_OBJECTS=$(TEST_ESP32_SRC:.cpp=.o)
@@ -52,7 +49,7 @@ TEST_JLED_GROUP_SRC=test_jled_group.cpp test_main.cpp ../src/jled_base.cpp ${CAT
 TEST_JLED_GROUP_OBJECTS=$(TEST_JLED_GROUP_SRC:.cpp=.o)
 
 all: bin bin/test_arduino_mock  \
-	 bin/test_jled bin/test_jled_sequence \
+	 bin/test_jled \
 	 bin/test_esp32_hal bin/test_arduino_hal bin/test_mbed_hal \
 	 bin/test_pico_hal bin/test_jled_esp8266_hal \
 	 bin/test_example_morse bin/test_brightness bin/test_fadeon_func \
@@ -63,9 +60,6 @@ bin/test_arduino_mock: $(TEST_ARDUINO_MOCK_OBJECTS)
 
 bin/test_jled: $(TEST_JLED_OBJECTS)
 	$(CXX) -o $@ $(LDFLAGS) $(TEST_JLED_OBJECTS)
-
-bin/test_jled_sequence: $(TEST_JLED_SEQUENCE_OBJECTS)
-	$(CXX) -o $@ $(LDFLAGS) $(TEST_JLED_SEQUENCE_OBJECTS)
 
 bin/test_esp32_hal: CFLAGS += -DESP32
 bin/test_esp32_hal: $(TEST_ESP32_OBJECTS)
@@ -104,7 +98,6 @@ coverage: test
 
 test: depend all
 	./bin/test_jled
-	./bin/test_jled_sequence
 	./bin/test_mbed_hal
 	./bin/test_esp32_hal
 	./bin/test_arduino_hal
@@ -133,7 +126,7 @@ clobber: clean
 
 depend: .depend
 
-(??).depend: $(TEST_JLED_SRC) $(TEST_JLED_SEQUENCE_SRC) $(TEST_ESP32_SRC) $(TEST_ESP8266_HAL_SRC) $(TEST_ARDUINO_SRC) $(TEST_MBED_SRC) $(TEST_PICO_SRC) $(TEST_MORSE_SRC) $(TEST_BRIGHTNESS_SRC) $(TEST_FADEON_SRC)
+.depend: $(TEST_JLED_SRC) $(TEST_ESP32_SRC) $(TEST_ESP8266_HAL_SRC) $(TEST_ARDUINO_SRC) $(TEST_MBED_SRC) $(TEST_PICO_SRC) $(TEST_MORSE_SRC) $(TEST_BRIGHTNESS_SRC) $(TEST_FADEON_SRC) $(TEST_JLED_GROUP_SRC)
 	@echo updating dependencies in .depend
 	@rm -f ./.depend
 	@$(CC) -I ../src -I . -I ./pico-sdk -MM $^ > .depend
